Fix links & navigation with folded notifications
An exploration into the implementation of the delete feature uncovered
another small bug in how the navigation works with folded notifications.

The `eosd-mode-link' function now reads the input arguments correctly
with `&optional' instead of `&rest' and binds the right variable as the
`key' in the `keymap'.

The code that powers `eosd-mode-delete-notification-under-cursor' is
shared with the one behind `eosd-mode-notification-body-toggle' through
a macro called `eosd-mode-notification-find-boundaries' that takes a
function and applies it with pointers for respectively begin and end of
the notification under the text cursor.